#1050 Extract a common MappingMethod for the normal (non-enum / non-value) mapping methods that are used within MapStruct
